Migrate MsgPack5.H5 library and UnitTests project to h5
Had a namespace / assembly resolution issue (that I've not been able to create a min reproduce case for yet) that meant that I dropped the "MsgPack5.H5" prefix from the namespaces and aseembly names other projects.

Changed the UnitTests project to write results to the document body instead of the console.